move ender cover into own package
