+++ src/main/java/gregtech/common/covers/CoverBehaviors.java
@@ -8,6 +8,7 @@ import gregtech.api.items.metaitem.MetaItem.MetaValueItem;
 import gregtech.api.util.GTLog;
 import gregtech.client.renderer.texture.Textures;
 import gregtech.common.covers.detector.*;
+import gregtech.common.covers.ender.CoverEnderFluidLink;
 import gregtech.common.items.MetaItems;
 import gregtech.common.items.behaviors.CoverDigitalInterfaceWirelessPlaceBehaviour;
